String attributes now use core::string so are safer (and no 1k limit) but slower. Relative positioning and automatic resizing GUI elements through setAlignment, setMaxSize and setMinSize (see changes.txt). rect<f32> override for setRelativePosition. All elements may can avoid being clipped by using setNotClipped, this only applies to parents (clip rect is recursive). added getFilesystem to gui environment. added automatic resizing example to the gui editor. fixed bug I introduced in the win32 device, when another window grabs the focus while buttons are held down.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@429 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CLimitReadFile.cpp
@@ -52,7 +52,7 @@ s32 CLimitReadFile::read(void* buffer, u32 sizeToRead)
 	if (pos >= AreaEnd)
 		return 0;
 
-	if (pos + sizeToRead >= AreaEnd)
+	if (pos + (s32)sizeToRead >= AreaEnd)
 		sizeToRead = AreaEnd - pos;
 
 	return File->read(buffer, sizeToRead);
